+++ include/seqan3/core/parallel/detail/all.hpp
@@ -12,6 +12,6 @@
 
 #pragma once
 
-#include <seqan3/core/parallel/detail/latch.hpp>
+#include <seqan3/utility/parallel/detail/latch.hpp>
 #include <seqan3/utility/parallel/detail/reader_writer_manager.hpp>
 #include <seqan3/utility/parallel/detail/spin_delay.hpp>
