+++ include/seqan3/utility/parallel/detail/reader_writer_manager.hpp
@@ -17,9 +17,9 @@
 #include <mutex>
 #include <seqan3/std/new>
 
-#include <seqan3/core/parallel/detail/latch.hpp>
-#include <seqan3/utility/parallel/detail/spin_delay.hpp>
 #include <seqan3/core/detail/strong_type.hpp>
+#include <seqan3/utility/parallel/detail/latch.hpp>
+#include <seqan3/utility/parallel/detail/spin_delay.hpp>
 
 namespace seqan3::detail
 {
